Moved console to own file
